+++ SoftServeCarMobile/babel.config.js
@@ -2,7 +2,7 @@ module.exports = {
   plugins: [
     "babel-plugin-transform-typescript-metadata",
     ["@babel/plugin-proposal-decorators", { "legacy": true }],
-    ["@babel/plugin-proposal-class-properties", { "loose": true }],
+    // ["@babel/plugin-proposal-class-properties", { "loose": true }],
   ],
   presets: ['module:metro-react-native-babel-preset'],
 };
