+++ SoftServeCarMobile/src/components/сhat/Chat.tsx
@@ -1,6 +1,5 @@
 import React from 'react';
 import * as signalR from '@microsoft/signalr';
-import styles from './ChatStyles/ChatStyles';
 
 
 import {
@@ -8,6 +7,7 @@ import {
     Text,
     TextInput,
     Button,
+    StyleSheet,
 } from 'react-native';
 
 
@@ -79,3 +79,42 @@ class Chat extends React.Component<ChatState, ChatState>{
 }
 
 export default Chat;
+
+
+const styles = StyleSheet.create({
+    container: {
+        width: '100%',
+        flex: 1,
+        justifyContent: 'flex-end',
+        alignItems: 'center',
+        marginBottom: 20,
+      },
+     
+    message: {
+        borderWidth: 1,
+        borderColor: "#20232a",
+        borderRadius: 8,
+        textAlign: 'center',
+        padding: 8,
+    },
+    input: {
+        padding: 10,
+        width: '80%',
+        borderWidth: 4,
+        borderColor: "#20232a",
+        borderRadius: 4,
+      },
+
+    buttonContainer: {
+        display: 'flex',
+        alignItems: 'center',
+        margin: 5,
+        flexDirection: 'row',
+        height: 40,
+        justifyContent: 'space-between'
+    },
+    chatMessage: {
+        //flex: 1,
+        backgroundColor: 'powderblue'
+    },    
+});
